+++ src/lib/components/table/evm-transactions/EvmTransactionsTableRow.tsx
@@ -93,7 +93,8 @@ export const EvmTransactionsTableRow = ({
             {formatUTokenWithPrecision(
               token.amount,
               token.precision ?? 0,
-              true
+              true,
+              token.precision ? 6 : 0
             )}
           </Text>
           {getTokenLabel(token.denom, token.symbol)}